tests: fix definition of fixture to fix test_workbench
